requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy==1.22.3
pandas==1.4.2
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
schedule==1.1.0
six==1.16.0
SQLAlchemy==1.4.35
termcolor==1.1.0
toml==0.10.2
urllib3==1.26.9
websocket-client==1.3.2
Werkzeug==2.1.1
zipp==3.8.0
